requirements.txt
psycopg2==2.9.10
Django>=4.2
djangorestframework==3.14.0
gunicorn
django-cors-headers>=4.0.0
TfidfTransformer>=0.1
cosine_similarity>=0.1.2
pandas>=2.2.3
google-generativeai
openpyxl